refactor: move cli commands to separate module
